th06nc: add multi-version support & support for latest update
- add HotKey constructor param to skip setup for runtime-setup hotkeys
  (addr not known compile-time)
- fix misc compiler warnings

+++ thprac_common/src/thprac_hook.cpp
@@ -15,7 +15,7 @@ void HookCtx::PatchSwap()
 {
     DWORD oldProt;
     VirtualProtect((LPVOID)this->addr, this->data.buffer.size, PAGE_EXECUTE_READWRITE, &oldProt);
-    memswap((void*)this->addr, this->data.buffer.ptr, this->data.buffer.size);
+    memswap((void*)this->addr, this->data.buffer.ptr, (unsigned int)this->data.buffer.size);
     VirtualProtect((LPVOID)this->addr, this->data.buffer.size, oldProt, &oldProt);
 }
